+++ crates/node/src/tests/wasm/snark.rs
@@ -8,7 +8,7 @@ use crate::backend::snark::Input;
 use wasm_bindgen_test::*;
 use crate::prelude::rings_core::utils::js_utils;
 use wasm_bindgen_futures::JsFuture;
-
+use super::setup_log;
 wasm_bindgen_test_configure!(run_in_browser);
 
 #[wasm_bindgen_test]
@@ -31,6 +31,23 @@ fn test_map_array_to_input() {
 
 #[wasm_bindgen_test]
 async fn test_send_snark_backend_message() {
+    setup_log();
+    let wasm = "https://raw.githubusercontent.com/RingsNetwork/rings/master/crates/snark/src/tests/circoms/simple_bn256.wasm";
+    let r1cs = "https://raw.githubusercontent.com/RingsNetwork/rings/master/crates/snark/src/tests/circoms/simple_bn256.r1cs";
+
+    let snark_behaviour = crate::backend::snark::SNARKBehaviour::new_instance();
+    let snark_task_builder = crate::backend::snark::SNARKTaskBuilder::from_remote(
+	r1cs.to_string(),
+	wasm.to_string(),
+	crate::backend::snark::SupportedPrimeField::Vesta
+    ).await.unwrap();
+    type F = crate::backend::snark::Field;
+    let input: Input = vec![
+	("step_in".to_string(), vec![F::from_u64(4u64, SupportedPrimeField::Vesta), F::from_u64(2u64, SupportedPrimeField::Vesta)])
+    ].into();
+    console_log!("gen circuit");
+    let circuits = snark_task_builder.gen_circuits(input, vec![], 5).unwrap();
+
     let provider1 = super::new_provider().await;
     let provider2 = super::new_provider().await;
 
@@ -39,29 +56,11 @@ async fn test_send_snark_backend_message() {
         JsFuture::from(provider2.listen()),
     )
 	.unwrap();
-    // let wasm = "https://github.com/RingsNetwork/rings/raw/master/crates/snark/src/tests/circoms/simple_bn256.wasm";
-    // let r1cs = "https://github.com/RingsNetwork/rings/raw/master/crates/snark/src/tests/circoms/simple_bn256.r1cs";
 
     super::create_connection(&provider1, &provider2).await;
     console_log!("wait for register");
     js_utils::window_sleep(1000).await.unwrap();
-
-    let r1cs = "http://localhost:7777/merkle_tree.r1cs";
-    let wasm = "http://localhost:7777/merkle_tree.wasm";
-
-    let snark_behaviour = crate::backend::snark::SNARKBehaviour::new_instance();
-    let snark_task_builder = crate::backend::snark::SNARKTaskBuilder::from_remote(
-	r1cs.to_string(),
-	wasm.to_string(),
-	crate::backend::snark::SupportedPrimeField::Pallas
-    ).await.unwrap();
-    type F = crate::backend::snark::Field;
-    let input: Input = vec![
-	("leaf".to_string(), vec![F::from_u64(4u64, SupportedPrimeField::Pallas)])
-    ].into();
-    console_log!("gen circuit");
-    let circuits = snark_task_builder.gen_circuits(input, vec![], 5).unwrap();
-
+    console_log!("gen snark task and send");
     let promise = snark_behaviour.send_proof_task_to(provider1, circuits, provider2.address());
     wasm_bindgen_futures::JsFuture::from(promise).await.unwrap();
 }
